+++ src/judge/note/touch.rs
@@ -5,7 +5,7 @@ use crate::materialize::MaterializedTouch;
 #[derive(Clone, Debug)]
 pub struct Touch {
     pub sensor: TouchSensor,
-    pub appear_time: f32,
+    pub appear_time: f64,
 
     judge_type: JudgeType,
 
@@ -24,11 +24,11 @@ impl From<MaterializedTouch> for Touch {
 }
 
 impl JudgeNote for Touch {
-    fn get_start_time(&self) -> f32 {
+    fn get_start_time(&self) -> f64 {
         self.appear_time + JUDGE_DATA.judge_param(self.judge_type).as_ref()[Timing::TooFast]
     }
 
-    fn get_end_time(&self) -> f32 {
+    fn get_end_time(&self) -> f64 {
         self.appear_time + JUDGE_DATA.judge_param(self.judge_type).as_ref()[Timing::LateGood]
     }
 
@@ -36,7 +36,7 @@ impl JudgeNote for Touch {
         Some(self.sensor)
     }
 
-    fn on_sensor(&mut self, current_time: f32) -> OnSensorResult {
+    fn on_sensor(&mut self, current_time: f64) -> OnSensorResult {
         assert!(self.result.is_none());
         if self.is_too_fast(current_time) {
             return OnSensorResult::TooFast;
@@ -49,7 +49,7 @@ impl JudgeNote for Touch {
         }
     }
 
-    fn judge(&mut self, _simulator: &TouchSensorStates, current_time: f32) {
+    fn judge(&mut self, _simulator: &TouchSensorStates, current_time: f64) {
         assert!(self.result.is_none());
         if self.is_too_late(current_time) {
             self.result = Some(Timing::TooLate);
